Reorganize stats components
