+++ services/main-frontend/src/components/page-specific/manage/courses/id/stats/CourseStatsPage.tsx
@@ -6,30 +6,30 @@ import { useTranslation } from "react-i18next"
 
 import { CourseManagementPagesProps } from "../../../../../../pages/manage/courses/[id]/[...path]"
 
-import AverageTimeToSubmit from "./AverageTimeToSubmit"
-import CourseSubmissionsByDay from "./CourseSubmissionsByDay"
-import CourseSubmissionsByWeekdayAndHour from "./CourseSubmissionsByWeekdayAndHour"
-import CourseUsersCountsByExercise from "./CourseUsersCountsByExercise"
-import CourseUsersWithSubmissionsByDay from "./CourseUsersWithSubmissionsByDay"
-import CourseVisitorsByCountry from "./CourseVisitorsByCountry"
-import CourseVisitorsByDay from "./CourseVisitorsByDay"
-import DailyCohortProgress from "./DailyCohortProgress"
-import DailyCompletions from "./DailyCompletions"
-import DailyUniqueUsers from "./DailyUniqueUsers"
-import DailyVisitCountsGroupedByReferrer from "./DailyVisitCountsGroupedByReferrer"
-import DailyVisitCountsGroupedByUtm from "./DailyVisitCountsGroupedByUtm"
-import DeviceTypes from "./DeviceTypes"
-import FirstSubmissionTrends from "./FirstSubmissionTrends"
-import MonthlyCompletions from "./MonthlyCompletions"
-import MonthlyReturningUsers from "./MonthlyReturningUsers"
-import MonthlyUniqueUsers from "./MonthlyUniqueUsers"
-import MostVisitedPages from "./MostVisitedPages"
-import TopReferrers from "./TopReferrers"
-import TopUtmCampaigns from "./TopUtmCampaigns"
-import TopUtmSources from "./TopUtmSources"
-import TotalStats from "./TotalStats"
-import WeeklyCohortProgress from "./WeeklyCohortProgress"
-import WeeklyUniqueUsers from "./WeeklyUniqueUsers"
+import CourseUsersCountsByExercise from "./visualizations/overview/CourseUsersCountsByExercise"
+import DailyCompletions from "./visualizations/overview/DailyCompletions"
+import MonthlyCompletions from "./visualizations/overview/MonthlyCompletions"
+import TotalStats from "./visualizations/overview/TotalStats"
+import AverageTimeToSubmit from "./visualizations/user-activity/AverageTimeToSubmit"
+import CourseSubmissionsByDay from "./visualizations/user-activity/CourseSubmissionsByDay"
+import CourseSubmissionsByWeekdayAndHour from "./visualizations/user-activity/CourseSubmissionsByWeekdayAndHour"
+import CourseUsersWithSubmissionsByDay from "./visualizations/user-activity/CourseUsersWithSubmissionsByDay"
+import DailyCohortProgress from "./visualizations/user-activity/DailyCohortProgress"
+import DailyUniqueUsers from "./visualizations/user-activity/DailyUniqueUsers"
+import FirstSubmissionTrends from "./visualizations/user-activity/FirstSubmissionTrends"
+import MonthlyReturningUsers from "./visualizations/user-activity/MonthlyReturningUsers"
+import MonthlyUniqueUsers from "./visualizations/user-activity/MonthlyUniqueUsers"
+import WeeklyCohortProgress from "./visualizations/user-activity/WeeklyCohortProgress"
+import WeeklyUniqueUsers from "./visualizations/user-activity/WeeklyUniqueUsers"
+import CourseVisitorsByCountry from "./visualizations/visitors/CourseVisitorsByCountry"
+import CourseVisitorsByDay from "./visualizations/visitors/CourseVisitorsByDay"
+import DailyVisitCountsGroupedByReferrer from "./visualizations/visitors/DailyVisitCountsGroupedByReferrer"
+import DailyVisitCountsGroupedByUtm from "./visualizations/visitors/DailyVisitCountsGroupedByUtm"
+import DeviceTypes from "./visualizations/visitors/DeviceTypes"
+import MostVisitedPages from "./visualizations/visitors/MostVisitedPages"
+import TopReferrers from "./visualizations/visitors/TopReferrers"
+import TopUtmCampaigns from "./visualizations/visitors/TopUtmCampaigns"
+import TopUtmSources from "./visualizations/visitors/TopUtmSources"
 
 import TabLink from "@/shared-module/common/components/Navigation/TabLinks/TabLink"
 import TabLinkNavigation from "@/shared-module/common/components/Navigation/TabLinks/TabLinkNavigation"
