+++ services/main-frontend/src/components/page-specific/manage/courses/id/stats/visualizations/overview/TotalStats.tsx
