+++ services/main-frontend/src/components/page-specific/manage/courses/id/stats/visualizations/user-activity/DailyCohortProgress.tsx
