+++ services/main-frontend/src/components/page-specific/manage/courses/id/stats/visualizations/visitors/CourseVisitorsByCountry.tsx
@@ -1,9 +1,10 @@
 import { css } from "@emotion/css"
+import { useQuery } from "@tanstack/react-query"
 import React, { useMemo } from "react"
 
-import useCoursePageVisitDatumSummary from "../../../../../../hooks/useCoursePageVisitDatumSummary"
+import { fetchCoursePageVisitDatumSummariesByCountry } from "@/services/backend/courses"
 
-import Echarts from "./Echarts"
+import Echarts from "../../Echarts"
 
 import DebugModal from "@/shared-module/common/components/DebugModal"
 import ErrorBanner from "@/shared-module/common/components/ErrorBanner"
@@ -12,30 +13,44 @@ import { baseTheme } from "@/shared-module/common/styles"
 import { dontRenderUntilQueryParametersReady } from "@/shared-module/common/utils/dontRenderUntilQueryParametersReady"
 import withErrorBoundary from "@/shared-module/common/utils/withErrorBoundary"
 
-export interface TopUTMCampaignsProps {
+export interface CourseVisitorsByCountryProps {
   courseId: string
 }
 
-const TopUTMCampaigns: React.FC<React.PropsWithChildren<TopUTMCampaignsProps>> = ({ courseId }) => {
-  const query = useCoursePageVisitDatumSummary(courseId)
+const CourseVisitorsByCountry: React.FC<React.PropsWithChildren<CourseVisitorsByCountryProps>> = ({
+  courseId,
+}) => {
+  const query = useQuery({
+    queryKey: [`course-page-visit-datum-summary-by-country${courseId}`],
+    queryFn: () => fetchCoursePageVisitDatumSummariesByCountry(courseId),
+  })
 
   const aggregatedData = useMemo(() => {
     if (!query.data || query.data.length === 0) {
       return null
     }
-    const allUtmCampaignsInData = Array.from(
-      new Set(query.data.map((item) => item.utm_campaign)),
-    ).filter((item) => !!item)
-    const totalCountsByUTMCampaign: { [referrer: string]: number } = Array.from(
-      allUtmCampaignsInData,
-    ).reduce((acc, utm_campaign) => {
-      const totalCount = query.data
-        .filter((item) => item.utm_campaign === utm_campaign)
-        .reduce((acc, item) => acc + item.num_visitors, 0)
-      // eslint-disable-next-line i18next/no-literal-string
-      return { ...acc, [utm_campaign ?? "null"]: totalCount }
-    }, {})
-    return totalCountsByUTMCampaign
+    const allCountriesInData = new Set(query.data.map((d) => d.country))
+    let totalCountsByCountry = Array.from(allCountriesInData)
+      .map((country) => {
+        const countryData = query.data.filter((d) => d.country === country)
+        return {
+          country,
+          num_visitors: countryData.reduce((acc, d) => acc + d.num_visitors, 0),
+        }
+      })
+      .sort((a, b) => a.num_visitors - b.num_visitors)
+    if (totalCountsByCountry.length > 15) {
+      totalCountsByCountry = totalCountsByCountry.filter((d) => d.num_visitors >= 10)
+    }
+    const totalCountsByCountryObject = totalCountsByCountry.reduce(
+      (acc, d) => {
+        // eslint-disable-next-line i18next/no-literal-string
+        acc[d.country ?? "null"] = d.num_visitors
+        return acc
+      },
+      {} as Record<string, number>,
+    )
+    return totalCountsByCountryObject
   }, [query.data])
 
   const categories = useMemo(() => {
@@ -77,10 +92,6 @@ const TopUTMCampaigns: React.FC<React.PropsWithChildren<TopUTMCampaignsProps>> =
           <Echarts
             height={200 + categories.length * 25}
             options={{
-              grid: {
-                containLabel: true,
-                left: 0,
-              },
               yAxis: {
                 type: "category",
                 data: categories,
@@ -109,4 +120,4 @@ const TopUTMCampaigns: React.FC<React.PropsWithChildren<TopUTMCampaignsProps>> =
   )
 }
 
-export default withErrorBoundary(dontRenderUntilQueryParametersReady(TopUTMCampaigns))
+export default withErrorBoundary(dontRenderUntilQueryParametersReady(CourseVisitorsByCountry))
