+++ services/main-frontend/src/components/page-specific/manage/courses/id/stats/visualizations/visitors/DeviceTypes.tsx
@@ -3,9 +3,9 @@ import { useQuery } from "@tanstack/react-query"
 import React, { useMemo } from "react"
 import { useTranslation } from "react-i18next"
 
-import { fetchCoursePageVisitDatumSummariesByDeviceTypes } from "../../../../../../services/backend/courses"
+import { fetchCoursePageVisitDatumSummariesByDeviceTypes } from "@/services/backend/courses"
 
-import Echarts from "./Echarts"
+import Echarts from "../../Echarts"
 
 import DebugModal from "@/shared-module/common/components/DebugModal"
 import ErrorBanner from "@/shared-module/common/components/ErrorBanner"
