+++ services/main-frontend/src/components/page-specific/manage/courses/id/stats/visualizations/visitors/MostVisitedPages.tsx
@@ -1,9 +1,11 @@
 import { css } from "@emotion/css"
+import { useQuery } from "@tanstack/react-query"
 import React, { useMemo } from "react"
 
-import useCoursePageVisitDatumSummary from "../../../../../../hooks/useCoursePageVisitDatumSummary"
+import { useCourseStructure } from "@/hooks/useCourseStructure"
+import { fetchCoursePageVisitDatumSummaryByPages } from "@/services/backend/courses"
 
-import Echarts from "./Echarts"
+import Echarts from "../../Echarts"
 
 import DebugModal from "@/shared-module/common/components/DebugModal"
 import ErrorBanner from "@/shared-module/common/components/ErrorBanner"
@@ -12,38 +14,53 @@ import { baseTheme } from "@/shared-module/common/styles"
 import { dontRenderUntilQueryParametersReady } from "@/shared-module/common/utils/dontRenderUntilQueryParametersReady"
 import withErrorBoundary from "@/shared-module/common/utils/withErrorBoundary"
 
-export interface TopUTMSourcesProps {
+export interface CourseVisitorsByCountryProps {
   courseId: string
 }
 
-const TopUTMSources: React.FC<React.PropsWithChildren<TopUTMSourcesProps>> = ({ courseId }) => {
-  const query = useCoursePageVisitDatumSummary(courseId)
+const CourseVisitorsByCountry: React.FC<React.PropsWithChildren<CourseVisitorsByCountryProps>> = ({
+  courseId,
+}) => {
+  const query = useQuery({
+    queryKey: [`course-page-visit-datum-summary-by-pages-${courseId}`],
+    queryFn: () => fetchCoursePageVisitDatumSummaryByPages(courseId),
+  })
+
+  const courseStructure = useCourseStructure(courseId)
 
   const aggregatedData = useMemo(() => {
     if (!query.data || query.data.length === 0) {
       return null
     }
-    const allUtmSourcesInData = Array.from(
-      new Set(query.data.map((item) => item.utm_source)),
-    ).filter((item) => !!item)
-    const totalCountsByUTMSource: { [referrer: string]: number } = Array.from(
-      allUtmSourcesInData,
-    ).reduce((acc, utm_source) => {
-      const totalCount = query.data
-        .filter((item) => item.utm_source === utm_source)
-        .reduce((acc, item) => acc + item.num_visitors, 0)
-      // eslint-disable-next-line i18next/no-literal-string
-      return { ...acc, [utm_source ?? "null"]: totalCount }
-    }, {})
-    return totalCountsByUTMSource
+    const allPageIdsInData = Array.from(new Set(query.data.map((obj) => obj.page_id)))
+    const totalCountsByPage = Array.from(allPageIdsInData)
+      .map((pageId) => {
+        const pageData = query.data.filter((item) => item.page_id === pageId)
+        return {
+          page_id: pageId,
+          total: pageData.reduce((acc, curr) => acc + curr.num_visitors, 0),
+        }
+      })
+      .sort((a, b) => a.total - b.total)
+    const topPages: { [page_id: string]: number } = totalCountsByPage
+      .slice(-100)
+      .reduce((acc, curr) => {
+        return {
+          ...acc,
+          [curr.page_id]: curr.total,
+        }
+      }, {})
+    return topPages
   }, [query.data])
 
   const categories = useMemo(() => {
     if (!aggregatedData) {
       return []
     }
-    return Object.keys(aggregatedData)
-  }, [aggregatedData])
+    return Object.keys(aggregatedData).map((pageId) => {
+      return courseStructure.data?.pages.find((page) => page.id === pageId)?.url_path ?? pageId
+    })
+  }, [aggregatedData, courseStructure.data?.pages])
   const values = useMemo(() => {
     if (!aggregatedData) {
       return []
@@ -55,7 +72,7 @@ const TopUTMSources: React.FC<React.PropsWithChildren<TopUTMSourcesProps>> = ({ 
     return <ErrorBanner variant="readOnly" error={query.error} />
   }
 
-  if (query.isPending || !query.data) {
+  if (query.isPending || !query.data || courseStructure.isPending) {
     return <Spinner variant="medium" />
   }
 
@@ -109,4 +126,4 @@ const TopUTMSources: React.FC<React.PropsWithChildren<TopUTMSourcesProps>> = ({ 
   )
 }
 
-export default withErrorBoundary(dontRenderUntilQueryParametersReady(TopUTMSources))
+export default withErrorBoundary(dontRenderUntilQueryParametersReady(CourseVisitorsByCountry))
